Extract InMemoryBusEngine into its own project
