+++ Nybus.sln
@@ -43,9 +43,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SenderFeedback", "samples\R
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Types", "samples\RabbitMQ\Types\Types.csproj", "{5EADDCA9-0316-4596-85B1-6C91EDFE8460}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests.Integration.Nybus", "tests\Tests.Integration.Nybus\Tests.Integration.Nybus.csproj", "{C498B46D-26EF-4AAE-ADFA-5763609A688B}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestUtils", "tests\TestUtils\TestUtils.csproj", "{6DAC88A4-9DCF-490B-81D3-B0AA72993EA4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nybus.Engine.InMemory", "src\engines\Nybus.Engine.InMemory\Nybus.Engine.InMemory.csproj", "{C59D94B8-9869-4E12-B393-2645E9C38BF3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests.Nybus.Engine.InMemory", "tests\Tests.Nybus.Engine.InMemory\Tests.Nybus.Engine.InMemory.csproj", "{8AA01AE0-1F81-41E4-917C-682E22B645D2}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests.Integration.Nybus.Engine.InMemory", "tests\Tests.Integration.Nybus.Engine.InMemory\Tests.Integration.Nybus.Engine.InMemory.csproj", "{6B36CB22-F186-4AD9-8A6D-13E7A3E6D3A2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -225,18 +229,6 @@ Global
 		{5EADDCA9-0316-4596-85B1-6C91EDFE8460}.Release|x64.Build.0 = Release|Any CPU
 		{5EADDCA9-0316-4596-85B1-6C91EDFE8460}.Release|x86.ActiveCfg = Release|Any CPU
 		{5EADDCA9-0316-4596-85B1-6C91EDFE8460}.Release|x86.Build.0 = Release|Any CPU
-		{C498B46D-26EF-4AAE-ADFA-5763609A688B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C498B46D-26EF-4AAE-ADFA-5763609A688B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C498B46D-26EF-4AAE-ADFA-5763609A688B}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{C498B46D-26EF-4AAE-ADFA-5763609A688B}.Debug|x64.Build.0 = Debug|Any CPU
-		{C498B46D-26EF-4AAE-ADFA-5763609A688B}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{C498B46D-26EF-4AAE-ADFA-5763609A688B}.Debug|x86.Build.0 = Debug|Any CPU
-		{C498B46D-26EF-4AAE-ADFA-5763609A688B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C498B46D-26EF-4AAE-ADFA-5763609A688B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C498B46D-26EF-4AAE-ADFA-5763609A688B}.Release|x64.ActiveCfg = Release|Any CPU
-		{C498B46D-26EF-4AAE-ADFA-5763609A688B}.Release|x64.Build.0 = Release|Any CPU
-		{C498B46D-26EF-4AAE-ADFA-5763609A688B}.Release|x86.ActiveCfg = Release|Any CPU
-		{C498B46D-26EF-4AAE-ADFA-5763609A688B}.Release|x86.Build.0 = Release|Any CPU
 		{6DAC88A4-9DCF-490B-81D3-B0AA72993EA4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{6DAC88A4-9DCF-490B-81D3-B0AA72993EA4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6DAC88A4-9DCF-490B-81D3-B0AA72993EA4}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -249,6 +241,42 @@ Global
 		{6DAC88A4-9DCF-490B-81D3-B0AA72993EA4}.Release|x64.Build.0 = Release|Any CPU
 		{6DAC88A4-9DCF-490B-81D3-B0AA72993EA4}.Release|x86.ActiveCfg = Release|Any CPU
 		{6DAC88A4-9DCF-490B-81D3-B0AA72993EA4}.Release|x86.Build.0 = Release|Any CPU
+		{C59D94B8-9869-4E12-B393-2645E9C38BF3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C59D94B8-9869-4E12-B393-2645E9C38BF3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C59D94B8-9869-4E12-B393-2645E9C38BF3}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C59D94B8-9869-4E12-B393-2645E9C38BF3}.Debug|x64.Build.0 = Debug|Any CPU
+		{C59D94B8-9869-4E12-B393-2645E9C38BF3}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C59D94B8-9869-4E12-B393-2645E9C38BF3}.Debug|x86.Build.0 = Debug|Any CPU
+		{C59D94B8-9869-4E12-B393-2645E9C38BF3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C59D94B8-9869-4E12-B393-2645E9C38BF3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C59D94B8-9869-4E12-B393-2645E9C38BF3}.Release|x64.ActiveCfg = Release|Any CPU
+		{C59D94B8-9869-4E12-B393-2645E9C38BF3}.Release|x64.Build.0 = Release|Any CPU
+		{C59D94B8-9869-4E12-B393-2645E9C38BF3}.Release|x86.ActiveCfg = Release|Any CPU
+		{C59D94B8-9869-4E12-B393-2645E9C38BF3}.Release|x86.Build.0 = Release|Any CPU
+		{8AA01AE0-1F81-41E4-917C-682E22B645D2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8AA01AE0-1F81-41E4-917C-682E22B645D2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8AA01AE0-1F81-41E4-917C-682E22B645D2}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{8AA01AE0-1F81-41E4-917C-682E22B645D2}.Debug|x64.Build.0 = Debug|Any CPU
+		{8AA01AE0-1F81-41E4-917C-682E22B645D2}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{8AA01AE0-1F81-41E4-917C-682E22B645D2}.Debug|x86.Build.0 = Debug|Any CPU
+		{8AA01AE0-1F81-41E4-917C-682E22B645D2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8AA01AE0-1F81-41E4-917C-682E22B645D2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8AA01AE0-1F81-41E4-917C-682E22B645D2}.Release|x64.ActiveCfg = Release|Any CPU
+		{8AA01AE0-1F81-41E4-917C-682E22B645D2}.Release|x64.Build.0 = Release|Any CPU
+		{8AA01AE0-1F81-41E4-917C-682E22B645D2}.Release|x86.ActiveCfg = Release|Any CPU
+		{8AA01AE0-1F81-41E4-917C-682E22B645D2}.Release|x86.Build.0 = Release|Any CPU
+		{6B36CB22-F186-4AD9-8A6D-13E7A3E6D3A2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6B36CB22-F186-4AD9-8A6D-13E7A3E6D3A2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6B36CB22-F186-4AD9-8A6D-13E7A3E6D3A2}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{6B36CB22-F186-4AD9-8A6D-13E7A3E6D3A2}.Debug|x64.Build.0 = Debug|Any CPU
+		{6B36CB22-F186-4AD9-8A6D-13E7A3E6D3A2}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6B36CB22-F186-4AD9-8A6D-13E7A3E6D3A2}.Debug|x86.Build.0 = Debug|Any CPU
+		{6B36CB22-F186-4AD9-8A6D-13E7A3E6D3A2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6B36CB22-F186-4AD9-8A6D-13E7A3E6D3A2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6B36CB22-F186-4AD9-8A6D-13E7A3E6D3A2}.Release|x64.ActiveCfg = Release|Any CPU
+		{6B36CB22-F186-4AD9-8A6D-13E7A3E6D3A2}.Release|x64.Build.0 = Release|Any CPU
+		{6B36CB22-F186-4AD9-8A6D-13E7A3E6D3A2}.Release|x86.ActiveCfg = Release|Any CPU
+		{6B36CB22-F186-4AD9-8A6D-13E7A3E6D3A2}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -271,8 +299,10 @@ Global
 		{05B81794-040C-49B4-A248-C72558582953} = {18B7CFF6-C18A-48FF-ADB2-557333ABCD6D}
 		{E2016F0E-AEB0-4D33-9456-33DCCBD41F09} = {18B7CFF6-C18A-48FF-ADB2-557333ABCD6D}
 		{5EADDCA9-0316-4596-85B1-6C91EDFE8460} = {18B7CFF6-C18A-48FF-ADB2-557333ABCD6D}
-		{C498B46D-26EF-4AAE-ADFA-5763609A688B} = {3566B926-8E32-4B0D-8DC0-6DB1EB82D2CB}
 		{6DAC88A4-9DCF-490B-81D3-B0AA72993EA4} = {3566B926-8E32-4B0D-8DC0-6DB1EB82D2CB}
+		{C59D94B8-9869-4E12-B393-2645E9C38BF3} = {22CE8614-4057-44E0-8140-FFB46C642539}
+		{8AA01AE0-1F81-41E4-917C-682E22B645D2} = {3566B926-8E32-4B0D-8DC0-6DB1EB82D2CB}
+		{6B36CB22-F186-4AD9-8A6D-13E7A3E6D3A2} = {3566B926-8E32-4B0D-8DC0-6DB1EB82D2CB}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {087BE4A7-5329-4F83-84DF-7AA56A3944EE}
