+++ samples/InMemory/NetCoreConsoleApp/NetCoreConsoleApp.csproj
@@ -12,6 +12,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\..\src\engines\Nybus.Engine.InMemory\Nybus.Engine.InMemory.csproj" />
     <ProjectReference Include="..\..\..\src\Nybus\Nybus.csproj" />
   </ItemGroup>
