+++ samples/InMemory/NetCoreConsoleApp/Program.cs
@@ -6,7 +6,6 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Configuration;
 using System.Collections.Generic;
 using System.Threading;
-using Nybus.Policies;
 using Nybus.Utils;
 
 namespace NetCoreConsoleApp
