+++ src/engines/Nybus.Engine.InMemory/InMemoryConfiguratorExtensions.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using Nybus.Configuration;
+
+namespace Nybus
+{
+    public static class InMemoryConfiguratorExtensions
+    {
+        public static void UseInMemoryBusEngine(this INybusConfigurator configurator)
+        {
+            configurator.UseBusEngine<InMemoryBusEngine>();
+        }
+    }
+}
