+++ src/engines/Nybus.Engine.InMemory/Nybus.Engine.InMemory.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <RootNamespace>Nybus</RootNamespace>
+    <AssemblyName>Nybus.Engine.InMemory</AssemblyName>
+    <Description />
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\Nybus.Abstractions\Nybus.Abstractions.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="System.Reactive.Linq" Version="3.1.1" />
+    <PackageReference Include="Kralizek.Extensions.Logging" Version="1.0.1" />
+  </ItemGroup>
+
+</Project>
