+++ tests/Tests.Integration.Nybus.Engine.InMemory/AutoMoqDataAttribute.cs
