+++ tests/Tests.Integration.Nybus.Engine.InMemory/AutomaticHandlerBareSetupTests.cs
