+++ tests/Tests.Integration.Nybus.Engine.InMemory/AutomaticHandlerConfigurationSetupTests.cs
