+++ tests/Tests.Integration.Nybus.Engine.InMemory/RegisteredHandlerBareSetupTests.cs
