+++ tests/Tests.Integration.Nybus.Engine.InMemory/SingletonHandlerBareSetupTests.cs
