+++ tests/Tests.Nybus.Engine.InMemory/AutoMoqDataAttribute.cs
@@ -1,0 +1,51 @@
+using System;
+using AutoFixture;
+using AutoFixture.AutoMoq;
+using AutoFixture.NUnit3;
+
+namespace Tests
+{
+    public class AutoMoqDataAttribute : AutoDataAttribute
+    {
+        public AutoMoqDataAttribute() : base(CreateFixture)
+        {
+            
+        }
+
+        private static IFixture CreateFixture()
+        {
+            IFixture fixture = new Fixture();
+
+            fixture.Customize(new AutoMoqCustomization
+            {
+                ConfigureMembers = true,
+                GenerateDelegates = true
+            });
+
+            fixture.Freeze<IServiceProvider>();
+
+            return fixture;
+        }
+    }
+
+    public class InlineAutoMoqDataAttribute : InlineAutoDataAttribute
+    {
+        public InlineAutoMoqDataAttribute(params object[] args) : base(CreateFixture, args)
+        {
+            
+        }
+
+        private static IFixture CreateFixture()
+        {
+            IFixture fixture = new Fixture();
+
+            fixture.Customize(new AutoMoqCustomization
+            {
+                ConfigureMembers = true,
+                GenerateDelegates = true
+            });
+
+            return fixture;
+        }
+    }
+}
